Move top-level errors to error module
There is a lot of error code in the daemon, this is the first step in
starting to clean it up.
